+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Ranamed onTrade to OnLevelClick. Reanmed TradeRequest to LevelClickEventArgs
+
 ## [0.1.7] - 2026-03-30
 
 ### Fixed